correct cast type
...

+++ temp/src/main/java/watchdog/temp/App.java
@@ -10,7 +10,7 @@ public class App
 
     public static void main(String[] args )
     {
-        int soc_temp = temp_check.temp_check();
+        float soc_temp = temp_check.temp_check();
         System.out.println(soc_temp);
     }
 }
